requirements.txt
boto3>=1.34.91, <2.0.0
botocore>=1.34.91, <2.0.0
click~=8.1.7
PyJWT>=2.8.0, <3.0.0
requests>=2.31.0, <3.0.0
setuptools>=49.2.1
chardet>=5.2.0, <6.0.0